+++ replica/database.cc
@@ -458,7 +458,7 @@ void backlog_controller::update_controller(float shares) {
 
 
 dirty_memory_manager::dirty_memory_manager(replica::database& db, size_t threshold, double soft_limit, scheduling_group deferred_work_sg)
-    : logalloc::region_group_reclaimer(threshold / 2, threshold * soft_limit / 2)
+    : dirty_memory_manager_logalloc::region_group_reclaimer(threshold / 2, threshold * soft_limit / 2)
     , _real_dirty_reclaimer(threshold)
     , _db(&db)
     , _real_region_group("memtable", _real_dirty_reclaimer, deferred_work_sg)
